refactor(codegen): decompose parts check by concern
