+++ scripts/codegen/src/plugins/parts/check/constants.ts
@@ -1,0 +1,73 @@
+// Valid JS identifier: starts with letter/underscore/$, followed by alphanumerics/_/$.
+export const JS_IDENTIFIER_RE = /^[A-Za-z_$][A-Za-z0-9_$]*$/;
+
+// Names that collide with codegen-emitted wrapper locals, HTML/React/Vue
+// attributes, or JS reserved words. Using one of these as the effective
+// propName produces a duplicate-identifier or parser error in the generated
+// React / Vue / contract code.
+export const RESERVED_PROP_NAMES = new Set([
+  // React / HTML / Vue conventions
+  "ref",
+  "className",
+  "class",
+  "children",
+  "key",
+  "style",
+  "id",
+  // Wrapper template locals — the React composite-list template emits
+  // `const { <propName> = [], className, ref, ...rest } = props;` followed by
+  // `const mergedClassName = …;` — any of these as a propName triggers a
+  // duplicate binding or shadowing.
+  "props",
+  "rest",
+  "mergedClassName",
+  // ES reserved words — using `default` / `let` / `case` / `class` etc. as a
+  // destructure key is a syntax error.
+  "break",
+  "case",
+  "catch",
+  "class",
+  "const",
+  "continue",
+  "debugger",
+  "default",
+  "delete",
+  "do",
+  "else",
+  "enum",
+  "export",
+  "extends",
+  "false",
+  "finally",
+  "for",
+  "function",
+  "if",
+  "import",
+  "in",
+  "instanceof",
+  "new",
+  "null",
+  "return",
+  "super",
+  "switch",
+  "this",
+  "throw",
+  "true",
+  "try",
+  "typeof",
+  "var",
+  "void",
+  "while",
+  "with",
+  "yield",
+  "let",
+  "implements",
+  "interface",
+  "package",
+  "private",
+  "protected",
+  "public",
+  "static",
+  "await",
+  "async",
+]);
