+++ scripts/codegen/src/plugins/parts/check/index.ts
@@ -1,0 +1,159 @@
+import type { Issue } from "../../../core/check-utils.ts";
+import { isComposite, issue } from "../../../core/check-utils.ts";
+import type { Spec, SpecPart } from "../../../schema.ts";
+import { checkNonRepeatingPart } from "./non-repeating-part.ts";
+import { checkRepeatingPart } from "./repeating-part.ts";
+import { checkGroupKeyPropCollisions, checkSiblingPropNameCollisions } from "./sibling-checks.ts";
+
+/**
+ * Rejections for repeating parts. Each rule has a stable number that
+ * doesn't change even when the wording shifts. Gaps in the numbering
+ * (e.g. 10 → 12) are intentional — rules retired or merged keep their
+ * slot empty so existing references stay valid.
+ *
+ *  1. `repeating: true` + `fromChildren: true` — contradictory.
+ *  2. `repeating: true` with no (or empty) `props:` — useless item shape.
+ *  3. Any part in a list composite declares nested `parts:` — deferred to #835.
+ *  4. Repeating part nested inside another repeating part — deferred to #834.
+ *  5. Two repeating siblings collapse to the same effective `propName` (and
+ *     are not legitimately sharing via `groupKey:`).
+ *  6. `propName:` and `groupKey:` both set on the same part — contradictory:
+ *     groupKey's value already supplies the shared prop name.
+ *  7. Two parts sharing `groupKey:` declare the same per-item prop name —
+ *     the merged item shape would have an ambiguous field.
+ *  8. Repeating part declares `props.id` — `id` is codegen-reserved.
+ *  9. A `groupKey:` value is referenced by exactly one repeating part —
+ *     `groupKey` is for shared-array siblings; use `propName:` for a single
+ *     part that wants a custom prop name.
+ * 10. The effective propName must be a valid JS identifier and must not
+ *     collide with codegen-emitted wrapper locals or JS reserved words (see
+ *     RESERVED_PROP_NAMES).
+ * 12. Repeating item props cannot set `responsive: true` — generators emit a
+ *     plain scalar field + single data-attr, no per-breakpoint expansion.
+ * 13. A list composite must declare exactly one non-repeating top-level part —
+ *     the wrapper. Multiple wrappers are silently dropped by the renderer;
+ *     zero wrappers cause a generator throw.
+ * 14. Repeating item prop names must be valid JS identifiers — codegen emits
+ *     `item.<name>` access in iteration bodies; hyphens / spaces / leading
+ *     digits produce parse errors.
+ * 15. `propName:` or `groupKey:` declared on a part without `repeating: true`
+ *     — both are only consumed when `repeating: true`. Silently ignored
+ *     downstream otherwise.
+ * 16. Group-level scalar prop on a non-repeating part in a list composite
+ *     uses an advanced shape the wrapper template can't currently handle.
+ *     Three sub-cases each emit their own issue: `responsive: true` (no
+ *     per-breakpoint expansion), `slot: true` (no slot body), and
+ *     `pattern: controllable` (no `default*` / `on*Change` triple).
+ */
+export function checkRepeatingParts(spec: Spec): Issue[] {
+  if (!isComposite(spec)) return [];
+  const issues: Issue[] = [];
+
+  // Walk the whole tree once to detect repeating presence. `anyRepeating`
+  // gates rule 3 (no nested parts in a list composite) and rule 13.
+  let anyRepeating = false;
+  const detect = (parts: Record<string, SpecPart>): void => {
+    for (const part of Object.values(parts)) {
+      if (part.repeating === true) anyRepeating = true;
+      if (part.parts) detect(part.parts);
+    }
+  };
+  detect(spec.parts);
+
+  // Rule 9: `groupKey:` value must be shared by ≥ 2 repeating parts across
+  // the spec. A lone groupKey is meaningless; the author wants `propName:`.
+  // Build a count and emit per-offender below if any group has size 1.
+  const groupKeyCounts = new Map<string, string[]>();
+  const collectGroupKeys = (parts: Record<string, SpecPart>): void => {
+    for (const [partName, part] of Object.entries(parts)) {
+      if (part.repeating === true && typeof part.groupKey === "string") {
+        const owners = groupKeyCounts.get(part.groupKey) ?? [];
+        owners.push(partName);
+        groupKeyCounts.set(part.groupKey, owners);
+      }
+      if (part.parts) collectGroupKeys(part.parts);
+    }
+  };
+  collectGroupKeys(spec.parts);
+  const loneGroupKeys = new Set<string>();
+  for (const [groupKey, owners] of groupKeyCounts) {
+    if (owners.length === 1) loneGroupKeys.add(groupKey);
+  }
+
+  // Rule 13: list composite has exactly one non-repeating top-level part.
+  // The composite-list renderers pick the first non-repeating top-level part
+  // as the wrapper; extras (or zero) cause silent drops or a generator throw.
+  if (anyRepeating) {
+    const topLevel = Object.entries(spec.parts);
+    const nonRepeatingTop = topLevel.filter(([, p]) => p.repeating !== true);
+    if (nonRepeatingTop.length === 0) {
+      issues.push(
+        issue(
+          spec.name,
+          "parts",
+          "list composite must declare exactly one non-repeating wrapper part — found 0.",
+        ),
+      );
+    } else if (nonRepeatingTop.length > 1) {
+      for (const [name] of nonRepeatingTop) {
+        issues.push(
+          issue(
+            spec.name,
+            `parts.${name}`,
+            `list composite must declare exactly one non-repeating wrapper part — found ${nonRepeatingTop.length}: ${nonRepeatingTop.map(([n]) => `'${n}'`).join(", ")}. Extra wrapper parts are silently dropped by the renderer.`,
+          ),
+        );
+      }
+    }
+  }
+
+  const walk = (
+    parts: Record<string, SpecPart>,
+    parentPath: string,
+    parentIsRepeating: boolean,
+  ): void => {
+    // Rules 5 and 7 — sibling-level collision checks.
+    checkSiblingPropNameCollisions(parts, parentPath, spec.name, issues);
+    checkGroupKeyPropCollisions(parts, parentPath, spec.name, issues);
+
+    for (const [partName, part] of Object.entries(parts)) {
+      const path = parentPath === "" ? `parts.${partName}` : `${parentPath}.parts.${partName}`;
+
+      // Rule 3 (generalized): in a list composite, no part — whether the
+      // repeating one or a non-repeating wrapper — may declare nested `parts:`.
+      // The composite-list renderers only emit the chosen wrapper + each
+      // repeating sibling; any nested sub-parts (under either) are silently
+      // dropped. Defer to #835 / a future phase.
+      if (anyRepeating && part.parts && Object.keys(part.parts).length > 0) {
+        issues.push(
+          issue(
+            spec.name,
+            path,
+            `'${partName}' cannot declare nested \`parts:\` in a list composite — nested fixed sub-parts are deferred to #835.`,
+          ),
+        );
+      }
+
+      if (part.repeating === true) {
+        checkRepeatingPart(
+          spec.name,
+          partName,
+          part,
+          path,
+          parentIsRepeating,
+          loneGroupKeys,
+          issues,
+        );
+      } else {
+        checkNonRepeatingPart(spec.name, partName, part, path, anyRepeating, issues);
+      }
+
+      if (part.parts) {
+        walk(part.parts, path, parentIsRepeating || part.repeating === true);
+      }
+    }
+  };
+
+  walk(spec.parts, "", false);
+  return issues;
+}
